finish list command
