+++ internal/cmd/list.go
@@ -2,15 +2,13 @@ package cmd
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
 	"strings"
 
 	"github.com/cli/cli/v2/pkg/iostreams"
 	"github.com/cli/cli/v2/utils"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	"github.com/stateful/runme/internal/document"
+	"github.com/stateful/runme/internal/project"
 	"github.com/stateful/runme/internal/shell"
 )
 
@@ -21,51 +19,61 @@ func listCmd() *cobra.Command {
 		Short:   "List available commands",
 		Long:    "Displays list of parsed command blocks, their name, number of commands in a block, and description from a given markdown file, such as README.md.",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			blocks := document.CodeBlocks{}
-			var err error
-			_, present := os.LookupEnv("EXPERIMENTAL_CLI")
-			if present {
-				e, files := findAllMarkdownFiles()
-				err = e
-
-				for _, file := range files {
-					fFileName = filepath.Base(file)
-					fChdir = filepath.Dir(file)
-					fileBlocks, e := getCodeBlocks()
-					err = e
-					blocks = append(blocks, fileBlocks...)
-				}
-			} else {
-				blocks, err = getCodeBlocks()
-			}
-
+			p, err := project.New(fChdir)
 			if err != nil {
 				return err
 			}
 
-			// TODO: this should be taken from cmd.
-			io := iostreams.System()
-			//lint:ignore SA1019 utils is deprecated but that's ok for now.
-			table := utils.NewTablePrinter(io)
-
-			// table header
-			table.AddField(strings.ToUpper("Name"), nil, nil)
-			table.AddField(strings.ToUpper("First Command"), nil, nil)
-			table.AddField(strings.ToUpper("# of Commands"), nil, nil)
-			table.AddField(strings.ToUpper("Description!!"), nil, nil)
-			table.EndRow()
-
-			for _, block := range blocks {
-				lines := block.Lines()
-
-				table.AddField(block.Name(), nil, nil)
-				table.AddField(shell.TryGetNonCommentLine(lines), nil, nil)
-				table.AddField(fmt.Sprintf("%d", len(shell.StripComments(lines))), nil, nil)
-				table.AddField(block.Intro(), nil, nil)
-				table.EndRow()
+			blocks := project.CodeBlocks{}
+			var blockError error
+			if isInExperimentalMode() {
+				blocks, blockError = p.GetAllCodeBlocks(fAllowUnknown)
+			} else {
+				codeBlocks, err := p.GetCodeBlocks(fFileName, fAllowUnknown)
+				blocks = append(blocks, &project.FileCodeBlocks{
+					FileName:   fFileName,
+					CodeBlocks: codeBlocks,
+				})
+				blockError = err
 			}
 
-			return errors.Wrap(table.Render(), "failed to render")
+			if blockError != nil {
+				return blockError
+			}
+
+			for _, block := range blocks {
+				if len(block.CodeBlocks) == 0 {
+					continue
+				}
+
+				fmt.Println(">>> Commands for", block.FileName)
+
+				// TODO: this should be taken from cmd.
+				io := iostreams.System()
+				//lint:ignore SA1019 utils is deprecated but that's ok for now.
+				table := utils.NewTablePrinter(io)
+
+				// table header
+				table.AddField(strings.ToUpper("Name"), nil, nil)
+				table.AddField(strings.ToUpper("First Command"), nil, nil)
+				table.AddField(strings.ToUpper("Description"), nil, nil)
+				table.EndRow()
+				for _, code := range block.CodeBlocks {
+					lines := code.Lines()
+					table.AddField(code.Name(), nil, nil)
+					table.AddField(shell.TryGetNonCommentLine(lines), nil, nil)
+					table.AddField(code.Intro(), nil, nil)
+					table.EndRow()
+				}
+
+				err := table.Render()
+				if err != nil {
+					return errors.Wrap(err, "failed to render")
+				}
+				fmt.Print("\n")
+			}
+
+			return nil
 		},
 	}
